refactor: QRModal 구조 개선 및 하위 컴포넌트 통합
